# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  gameCode: string;
  gameName: string;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  description: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface Denomination {
  id: number;
  gameCode: string;
  productCode: string;
  productName: string;
  price: number;
  originalPrice: number;
  sellingPrice: number;
  isActive: boolean;
}

export interface CreateOrderBody {
  gameCode: string;
  productCode: string;
  gameUserId: string;
  /** @nullable */
  gameServerId?: string | null;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
}

export interface Order {
  id: number;
  orderId: string;
  gameCode: string;
  gameName: string;
  productCode: string;
  productName: string;
  gameUserId: string;
  /** @nullable */
  gameServerId: string | null;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  amount: number;
  paymentStatus: string;
  digiflazzStatus: string;
  /** @nullable */
  midtransOrderId: string | null;
  /** @nullable */
  paymentUrl: string | null;
  /** @nullable */
  snapToken: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderWithPayment {
  order: Order;
  /** @nullable */
  snapToken: string | null;
  /** @nullable */
  paymentUrl: string | null;
}

export interface PopularGame {
  gameCode: string;
  gameName: string;
  orderCount: number;
}

export interface TransactionStats {
  totalOrders: number;
  successOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  popularGames: PopularGame[];
}

export interface CallbackResponse {
  success: boolean;
  message: string;
}

export type PaymentNotificationBody = { [key: string]: unknown };