# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Top Up Game Store API - integrated with Digiflazz and Midtrans
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: orders
    description: Order and payment operations
  - name: transactions
    description: Transaction tracking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all games/products
      description: Returns list of all available game top up products grouped by category
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{gameCode}/denominations:
    get:
      operationId: getDenominations
      tags: [products]
      summary: Get denominations for a specific game
      parameters:
        - name: gameCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of denomination options
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Denomination"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new top up order
      description: Creates an order and initiates payment via Midtrans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created with Midtrans payment token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithPayment"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/check/{referenceId}:
    get:
      operationId: checkOrderByReference
      tags: [orders]
      summary: Check order status by reference/transaction ID
      parameters:
        - name: referenceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions/stats:
    get:
      operationId: getTransactionStats
      tags: [transactions]
      summary: Get transaction statistics
      responses:
        "200":
          description: Transaction stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionStats"

  /payment/notification:
    post:
      operationId: paymentNotification
      tags: [orders]
      summary: Midtrans payment notification webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Notification processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallbackResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Product:
      type: object
      properties:
        id:
          type: integer
        gameCode:
          type: string
        gameName:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - gameCode
        - gameName
        - category
        - imageUrl
        - description
        - isActive
        - createdAt

    Denomination:
      type: object
      properties:
        id:
          type: integer
        gameCode:
          type: string
        productCode:
          type: string
        productName:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        sellingPrice:
          type: number
        isActive:
          type: boolean
      required:
        - id
        - gameCode
        - productCode
        - productName
        - price
        - originalPrice
        - sellingPrice
        - isActive

    CreateOrderBody:
      type: object
      properties:
        gameCode:
          type: string
        productCode:
          type: string
        gameUserId:
          type: string
        gameServerId:
          type: ["string", "null"]
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
      required:
        - gameCode
        - productCode
        - gameUserId
        - customerName
        - customerEmail
        - customerPhone

    Order:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: string
        gameCode:
          type: string
        gameName:
          type: string
        productCode:
          type: string
        productName:
          type: string
        gameUserId:
          type: string
        gameServerId:
          type: ["string", "null"]
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        amount:
          type: number
        paymentStatus:
          type: string
        digiflazzStatus:
          type: string
        midtransOrderId:
          type: ["string", "null"]
        paymentUrl:
          type: ["string", "null"]
        snapToken:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderId
        - gameCode
        - gameName
        - productCode
        - productName
        - gameUserId
        - gameServerId
        - customerName
        - customerEmail
        - customerPhone
        - amount
        - paymentStatus
        - digiflazzStatus
        - midtransOrderId
        - paymentUrl
        - snapToken
        - createdAt
        - updatedAt

    OrderWithPayment:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        snapToken:
          type: ["string", "null"]
        paymentUrl:
          type: ["string", "null"]
      required:
        - order
        - snapToken
        - paymentUrl

    TransactionStats:
      type: object
      properties:
        totalOrders:
          type: integer
        successOrders:
          type: integer
        pendingOrders:
          type: integer
        totalRevenue:
          type: number
        popularGames:
          type: array
          items:
            $ref: "#/components/schemas/PopularGame"
      required:
        - totalOrders
        - successOrders
        - pendingOrders
        - totalRevenue
        - popularGames

    PopularGame:
      type: object
      properties:
        gameCode:
          type: string
        gameName:
          type: string
        orderCount:
          type: integer
      required:
        - gameCode
        - gameName
        - orderCount

    CallbackResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message