# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns list of all available game top up products grouped by category
 * @summary Get all games/products
 */
export const GetProductsResponseItem = zod.object({
  id: zod.number(),
  gameCode: zod.string(),
  gameName: zod.string(),
  category: zod.string(),
  imageUrl: zod.string().nullable(),
  description: zod.string().nullable(),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});
export const GetProductsResponse = zod.array(GetProductsResponseItem);

/**
 * @summary Get denominations for a specific game
 */
export const GetDenominationsParams = zod.object({
  gameCode: zod.coerce.string(),
});

export const GetDenominationsResponseItem = zod.object({
  id: zod.number(),
  gameCode: zod.string(),
  productCode: zod.string(),
  productName: zod.string(),
  price: zod.number(),
  originalPrice: zod.number(),
  sellingPrice: zod.number(),
  isActive: zod.boolean(),
});
export const GetDenominationsResponse = zod.array(GetDenominationsResponseItem);

/**
 * Creates an order and initiates payment via Midtrans
 * @summary Create a new top up order
 */
export const CreateOrderBody = zod.object({
  gameCode: zod.string(),
  productCode: zod.string(),
  gameUserId: zod.string(),
  gameServerId: zod.string().nullish(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string(),
});

/**
 * @summary Get order by ID
 */
export const GetOrderParams = zod.object({
  orderId: zod.coerce.string(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  orderId: zod.string(),
  gameCode: zod.string(),
  gameName: zod.string(),
  productCode: zod.string(),
  productName: zod.string(),
  gameUserId: zod.string(),
  gameServerId: zod.string().nullable(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string(),
  amount: zod.number(),
  paymentStatus: zod.string(),
  digiflazzStatus: zod.string(),
  midtransOrderId: zod.string().nullable(),
  paymentUrl: zod.string().nullable(),
  snapToken: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Check order status by reference/transaction ID
 */
export const CheckOrderByReferenceParams = zod.object({
  referenceId: zod.coerce.string(),
});

export const CheckOrderByReferenceResponse = zod.object({
  id: zod.number(),
  orderId: zod.string(),
  gameCode: zod.string(),
  gameName: zod.string(),
  productCode: zod.string(),
  productName: zod.string(),
  gameUserId: zod.string(),
  gameServerId: zod.string().nullable(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string(),
  amount: zod.number(),
  paymentStatus: zod.string(),
  digiflazzStatus: zod.string(),
  midtransOrderId: zod.string().nullable(),
  paymentUrl: zod.string().nullable(),
  snapToken: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get transaction statistics
 */
export const GetTransactionStatsResponse = zod.object({
  totalOrders: zod.number(),
  successOrders: zod.number(),
  pendingOrders: zod.number(),
  totalRevenue: zod.number(),
  popularGames: zod.array(
    zod.object({
      gameCode: zod.string(),
      gameName: zod.string(),
      orderCount: zod.number(),
    }),
  ),
});

/**
 * @summary Midtrans payment notification webhook
 */
export const PaymentNotificationBody = zod.record(zod.string(), zod.unknown());

export const PaymentNotificationResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});