# lib/api-zod/src/generated/types/callbackResponse.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const CallbackResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

export type CallbackResponse = zod.input<typeof CallbackResponse>;