# lib/api-zod/src/generated/types/createOrderBody.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const CreateOrderBody = zod.object({
  gameCode: zod.string(),
  productCode: zod.string(),
  gameUserId: zod.string(),
  gameServerId: zod.string().nullish(),
  customerName: zod.string(),
  customerEmail: zod.string(),
  customerPhone: zod.string(),
});

export type CreateOrderBody = zod.input<typeof CreateOrderBody>;