# lib/api-zod/src/generated/types/denomination.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const Denomination = zod.object({
  id: zod.number(),
  gameCode: zod.string(),
  productCode: zod.string(),
  productName: zod.string(),
  price: zod.number(),
  originalPrice: zod.number(),
  sellingPrice: zod.number(),
  isActive: zod.boolean(),
});

export type Denomination = zod.input<typeof Denomination>;