# lib/api-zod/src/generated/types/errorResponse.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const ErrorResponse = zod.object({
  error: zod.string(),
});

export type ErrorResponse = zod.input<typeof ErrorResponse>;