# lib/api-zod/src/generated/types/orderWithPayment.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const OrderWithPayment = zod.object({
  order: zod.object({
    id: zod.number(),
    orderId: zod.string(),
    gameCode: zod.string(),
    gameName: zod.string(),
    productCode: zod.string(),
    productName: zod.string(),
    gameUserId: zod.string(),
    gameServerId: zod.string().nullable(),
    customerName: zod.string(),
    customerEmail: zod.string(),
    customerPhone: zod.string(),
    amount: zod.number(),
    paymentStatus: zod.string(),
    digiflazzStatus: zod.string(),
    midtransOrderId: zod.string().nullable(),
    paymentUrl: zod.string().nullable(),
    snapToken: zod.string().nullable(),
    createdAt: zod.string(),
    updatedAt: zod.string(),
  }),
  snapToken: zod.string().nullable(),
  paymentUrl: zod.string().nullable(),
});

export type OrderWithPayment = zod.input<typeof OrderWithPayment>;