# lib/api-zod/src/generated/types/popularGame.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const PopularGame = zod.object({
  gameCode: zod.string(),
  gameName: zod.string(),
  orderCount: zod.number(),
});

export type PopularGame = zod.input<typeof PopularGame>;