# lib/api-zod/src/generated/types/product.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const Product = zod.object({
  id: zod.number(),
  gameCode: zod.string(),
  gameName: zod.string(),
  category: zod.string(),
  imageUrl: zod.string().nullable(),
  description: zod.string().nullable(),
  isActive: zod.boolean(),
  createdAt: zod.string(),
});

export type Product = zod.input<typeof Product>;