# lib/api-zod/src/generated/types/transactionStats.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Top Up Game Store API - integrated with Digiflazz and Midtrans
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const TransactionStats = zod.object({
  totalOrders: zod.number(),
  successOrders: zod.number(),
  pendingOrders: zod.number(),
  totalRevenue: zod.number(),
  popularGames: zod.array(
    zod.object({
      gameCode: zod.string(),
      gameName: zod.string(),
      orderCount: zod.number(),
    }),
  ),
});

export type TransactionStats = zod.input<typeof TransactionStats>;